+++ src/wrench/services/compute/virtualized_cluster/VirtualizedClusterService.cpp
@@ -43,7 +43,7 @@ namespace wrench {
                                                          bool supports_pilot_jobs,
                                                          std::vector<std::string> &execution_hosts,
                                                          std::map<std::string, std::string> plist,
-                                                         int scratch_size) :
+                                                         double scratch_size) :
             ComputeService(hostname, "virtualized_cluster_service", "virtualized_cluster_service",
                            supports_standard_jobs, supports_pilot_jobs,
                            scratch_size) {
